refactor: separate web-ui components on order and scope
