+++ packages/web-ui/src/components/charts/PageViewsBarChart.tsx
