+++ packages/web-ui/src/components/charts/SessionsChart.tsx
