+++ packages/web-ui/src/components/charts/index.ts
@@ -1,0 +1,4 @@
+// Chart components - Data visualization and analytics charts
+export { default as ChartUserByCountry } from './ChartUserByCountry';
+export { default as PageViewsBarChart } from './PageViewsBarChart';
+export { default as SessionsChart } from './SessionsChart';
