+++ packages/web-ui/src/components/common/index.ts
@@ -1,0 +1,4 @@
+// Common/reusable components - Shared UI components used across different features
+export { default as HighlightedCard } from './HighlightedCard';
+export { default as StatCard } from './StatCard';
+export type { StatCardProps } from './StatCard';
