+++ packages/web-ui/src/components/grids/index.ts
@@ -1,0 +1,3 @@
+// Grid components - Data tables and grid layouts
+export { default as CustomizedDataGrid } from './CustomizedDataGrid';
+export { renderAvatar, columns, rows } from './GridData';
