--- packages/web-ui/src/components/index.tsx
@@ -1,27 +1,0 @@
-// Components
-
-import AppNavbar from './AppNavbar';
-export { AppNavbar };
-
-import Header from './Header';
-export { Header };
-
-import MainGrid from './MainGrid';
-export { MainGrid };
-
-import SideMenu from './SideMenu';
-export { SideMenu };
-
-// Shared
-// export { } from '';
-
-// Import for re-export object
-// import { } from '';
-
-// Re-export all components as a single object for easier importing
-export const Components = {
-  AppNavbar,
-  Header,
-  MainGrid,
-  SideMenu,
-};
