+++ packages/web-ui/src/components/pages/index.ts
@@ -1,0 +1,2 @@
+// Page-level components - Complete page layouts and page-specific components
+export { default as MainGrid } from './MainGrid';
