+++ packages/web-ui/src/components/shell/index.ts
@@ -1,0 +1,9 @@
+// Shell components - Navigation, menus, and shell-level UI elements
+export { default as AppNavbar } from './AppNavbar';
+export { default as Header } from './Header';
+export { default as SideMenu } from './SideMenu';
+export { default as SideMenuMobile } from './SideMenuMobile';
+export { default as MenuButton } from './MenuButton';
+export { default as MenuContent } from './MenuContent';
+export { default as NavbarBreadcrumbs } from './NavbarBreadcrumbs';
+export { default as OptionsMenu } from './OptionsMenu';
